update: upload image laporan dan konten header
-integrasi upload image dan submit laporan aman semua
-membuat redux untuk konten dinamis header
