+++ src/redux/store.js
@@ -1,11 +1,11 @@
 import {configureStore} from "@reduxjs/toolkit"
 import authReducers from "./reducers/auth-reducers"
-import laporanReducers from "./reducers/laporan-reducers"
+import profilReducers from "./reducers/profil-reducer"
 
 const store = configureStore({
     reducer:{
         authUser: authReducers,
-        laporan: laporanReducers
+        profilUser: profilReducers
     }
 })
